+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates are in YYYY-MM-DD format.
 - Adds raw options to BuildFlags
 - Adds `+` and `+=` overloads for BuildFlags
 - Removes all node categories except `Node`, which now only keeps track of arbitrary paths and associated commands
+- Adds Project, which can track one or more directories and the files contained within.
 
 ## SRBuild v0.1.0 (2019-02-16)
 - Initial version, basic compiler functionality.